Feat: 지역별 카테고리 선택 기능 구현
