+++ src/constants/districtTab.ts
@@ -1,0 +1,90 @@
+import { COLORS } from '../styles/root'
+
+const districtTab = [
+  {
+    categoryId: 9,
+    district: '동남아시아',
+    color: COLORS.c813333,
+    bgColor: COLORS.cFFB0B0,
+  },
+  {
+    categoryId: 10,
+    district: '괌&사이판&하와이',
+    color: COLORS.c813346,
+    bgColor: COLORS.cFFB0E9,
+  },
+  {
+    categoryId: 11,
+    district: '호주&뉴질랜드',
+    color: COLORS.c5E3381,
+    bgColor: COLORS.cDCB0FF,
+  },
+  {
+    categoryId: 13,
+    district: '인도&주변국',
+    color: COLORS.c335781,
+    bgColor: COLORS.cb0e3ff,
+  },
+  {
+    categoryId: 14,
+    district: '중앙아시아',
+    bgColor: COLORS.cffb5b0,
+    color: COLORS.c813333,
+  },
+  {
+    categoryId: 16,
+    district: '동남아프리카',
+    bgColor: COLORS.cfdffb0,
+    color: COLORS.c74642c,
+  },
+  {
+    categoryId: 17,
+    district: '북아프리카&중동',
+    bgColor: COLORS.cb0ffcb,
+    color: COLORS.c338136,
+  },
+  {
+    categoryId: 19,
+    district: '코카서스',
+    bgColor: COLORS.c74fff7,
+    color: COLORS.c1B5754,
+  },
+  {
+    categoryId: 20,
+    district: '유럽',
+    bgColor: COLORS.cFFE9B0,
+    color: COLORS.c725C24,
+  },
+  {
+    categoryId: 22,
+    district: '중남미',
+    bgColor: COLORS.cB0C6FF,
+    color: COLORS.c335281,
+  },
+  {
+    categoryId: 23,
+    district: '북미',
+    bgColor: COLORS.cE6FFB0,
+    color: COLORS.c3E8133,
+  },
+  {
+    categoryId: 25,
+    district: '대만',
+    bgColor: COLORS.cf39c5d,
+    color: COLORS.c563a20,
+  },
+  {
+    categoryId: 26,
+    district: '중국',
+    bgColor: COLORS.ca0ca8c,
+    color: COLORS.c324b26,
+  },
+  {
+    categoryId: 27,
+    district: '일본',
+    bgColor: COLORS.cb6adeb,
+    color: COLORS.c382e74,
+  },
+]
+
+export default districtTab
